Split the highlights locationChange reducer in 2 to fix disappearing highlights
