+++ src/app/content/highlights/hooks/loadHighlights.ts
